referal_system/requirements.txt
-i https://pypi.org/simple
asgiref==3.9.1; python_version >= '3.9'
attrs==25.3.0; python_version >= '3.8'
django==5.2.4; python_version >= '3.10'
djangorestframework==3.16.0; python_version >= '3.9'
djangorestframework-simplejwt==5.5.1; python_version >= '3.9'
drf-spectacular[sidecar]==0.28.0; python_version >= '3.7'
drf-spectacular-sidecar==2025.7.1; python_version >= '3.6'
inflection==0.5.1; python_version >= '3.5'
jsonschema==4.25.0; python_version >= '3.9'
jsonschema-specifications==2025.4.1; python_version >= '3.9'
pyjwt==2.10.1; python_version >= '3.9'
pyyaml==6.0.2; python_version >= '3.8'
referencing==0.36.2; python_version >= '3.9'
rpds-py==0.26.0; python_version >= '3.9'
sqlparse==0.5.3; python_version >= '3.8'
typing-extensions==4.14.1; python_version >= '3.9'
tzdata==2025.2; python_version >= '2'
uritemplate==4.2.0; python_version >= '3.9'
